pyproject.toml
[tool.poetry]
name = "bel_resources"
version = "0.2.0"
description = "BEL Resource Generator - namespaces, orthologies, etc"
authors = ["William Hayes <william.s.hayes@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
bel = "*"
certifi = "*"
chardet = "*"
click = "*"
elasticsearch = "*"
fastcache = "*"
idna = "*"
pronto = "*"
pytest = "*"
pytest-cache = "*"
pytest-cov = "*"
pytest-flakes = "*"
pytest-mypy = "*"
"pytest-pep8" = "*"
pytest-sugar = "*"
python-json-logger = "*"
python-dateutil = "*"
requests = "*"
six = "*"
structlog = "*"
timy = "*"
tornado = "*"
"urllib3" = "*"
PyYAML = "*"
python-arango = "^4.4"
python-dotenv = "^0.10.3"
flake8 = "^3.7.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
